+++ src/main/java/com/nttdatacasefirst/stockAPI/dtos/CapitalIncreaseGetModel.java
@@ -19,7 +19,7 @@ public class CapitalIncreaseGetModel {
     private BigDecimal rightIssue;
     @Digits(integer = 15, fraction = 2)
     private BigDecimal bonusIssue;
-    private byte increasingRate;
+    private int increasingRate;
     @Digits(integer = 15, fraction = 2)
     private BigDecimal capitalValue;
 }
